refactor: JobMotivationEmailPanel file name
