+++ web-manager/src/app/components/JobMotivationEmailPanel.tsx
